backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
SQLAlchemy==2.0.28
alembic==1.13.1
PyJWT==2.8.0
Werkzeug==3.0.1
python-dotenv==1.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
blinker==1.7.0